Added 2D texture support for shader skins & more. - Split skins into it's own script file. - Moved around some scripts for better structure. - Added lava skin.
