+++ public/resources/skins/lava/vertex.glsl
@@ -1,0 +1,7 @@
+varying vec2 vUv;
+void main()
+{
+	vUv=uv;
+	vec4 mvPosition=modelViewMatrix*vec4(position,1.);
+	gl_Position=projectionMatrix*mvPosition;
+}
